+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/Layer.java
@@ -218,7 +218,12 @@ public enum Layer {
     /**
      * A cloud native messaging and streaming platform, making it simple to build event-driven applications.
      */
-    ROCKETMQ(35, true);
+    ROCKETMQ(35, true),
+
+    /**
+     * A high-performance, column-oriented SQL database management system (DBMS) for online analytical processing (OLAP).
+     */
+    CLICKHOUSE(36, true);
 
     private final int value;
     /**