+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/management/ui/template/UITemplateInitializer.java
@@ -73,6 +73,7 @@ public class UITemplateInitializer {
         Layer.BOOKKEEPER.name(),
         Layer.NGINX.name(),
         Layer.ROCKETMQ.name(),
+        Layer.CLICKHOUSE.name(),
         "custom"
     };
     private final UITemplateManagementService uiTemplateManagementService;